refactor(engines): promote engine.py → engines/engine.py (slice 3/8, #170)
