+++ .importlinter
@@ -12,7 +12,7 @@ layers =
     voicecli.cli | voicecli.cli_samples | voicecli.cli_dictate | voicecli.cli_doctor | voicecli.cli_nats | voicecli.tg | voicecli.listen | voicecli.overlay | voicecli.stt_client | voicecli.overlay_audio | voicecli.overlay_draw
     voicecli.api
     voicecli.adapters
-    voicecli.engine | voicecli.transcribe | voicecli.daemon | voicecli.stt_daemon | voicecli.model_registry | voicecli.daemon_protocol
+    voicecli.engines.engine | voicecli.transcribe | voicecli.daemon | voicecli.stt_daemon | voicecli.model_registry | voicecli.daemon_protocol
     voicecli.config | voicecli.paths | voicecli.utils | voicecli.env | voicecli.models | voicecli.samples | voicecli.stt_modes | voicecli.clipboard | voicecli.history | voicecli.ui_sounds
 ignore_imports =
     # Transitional: voicecli.api.translate imports voicecli.api.markdown for MD rendering — same api package peer dependency
@@ -32,18 +32,18 @@ ignore_imports =
     voicecli.cli_nats -> voicecli.model_registry
     # Transitional: voicecli.cli imports voicecli.listen for recording subcommand — needs extraction to adapter layer
     voicecli.cli -> voicecli.listen
-    # Transitional: voicecli.daemon imports voicecli.engine to load TTS engines — daemon is an infrastructure adapter, coupling is expected but needs SynthesisPort wiring (ADR-003)
-    voicecli.daemon -> voicecli.engine
+    # Transitional: voicecli.daemon imports voicecli.engines.engine to load TTS engines — daemon is an infrastructure adapter, coupling is expected but needs SynthesisPort wiring (ADR-003)
+    voicecli.daemon -> voicecli.engines.engine
     # Transitional: daemon_protocol is a private helper of daemon — same infrastructure layer, peer import acceptable
     voicecli.daemon -> voicecli.daemon_protocol
     # Transitional: daemon_protocol directly loads engines — covered by the same SynthesisPort ADR-003 wiring work as daemon
-    voicecli.daemon_protocol -> voicecli.engine
+    voicecli.daemon_protocol -> voicecli.engines.engine
     # Transitional: daemon_protocol is shared wire-encoding between TTS/STT daemons (audit #162). Per ADR-003 debt #4, extract to voicecli.wire_protocol under the ports layer once layered axis materializes.
     voicecli.stt_daemon -> voicecli.daemon_protocol
     # Transitional: voicecli.stt_daemon imports voicecli.transcribe to run STT — coupling is expected but needs STTEnginePort wiring (ADR-003)
     voicecli.stt_daemon -> voicecli.transcribe
-    # Transitional: voicecli.model_registry imports voicecli.engine to manage engine instances — needs decoupling via ports layer (ADR-003)
-    voicecli.model_registry -> voicecli.engine
+    # Transitional: voicecli.model_registry imports voicecli.engines.engine to manage engine instances — needs decoupling via ports layer (ADR-003)
+    voicecli.model_registry -> voicecli.engines.engine
     # ADR-003 V10: same top-tier peer dependencies — overlay assembler imports extracted sub-modules
     voicecli.overlay -> voicecli.overlay_audio
     voicecli.overlay -> voicecli.overlay_draw
