+++ src/voicecli/engines/chatterbox.py
@@ -5,7 +5,7 @@ import soundfile as sf
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from voicecli.engine import cuda_guard
+from voicecli.engines.engine import cuda_guard
 from voicecli.engines._chatterbox_base import ChatterboxBase
 from voicecli.models import CHATTERBOX_MODEL, warn_if_first_download
 from voicecli.utils import resolve_language as _resolve_language
