+++ src/voicecli/engines/qwen.py
@@ -7,7 +7,7 @@ import torch
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from voicecli.engine import TTSEngine, cuda_guard
+from voicecli.engines.engine import TTSEngine, cuda_guard
 from voicecli.utils import to_full_language_name
 from voicecli.models import (
     QWEN_CLONE_MODEL,
